+++ lib/app/constants/app_config.dart
@@ -40,7 +40,7 @@ class AppConfig {
     //   }
     // }
 
-    // return 'http://bill.local:8000';
+     //return 'http://bill.local:8000';
     //  return 'http://192.168.31.83:8000';
      return 'https://bill.sakthicontroller.com/api/public';
   }